+++ __mocks__/http-proxy-middleware.js
@@ -1,0 +1,4 @@
+const middleware = () => {}
+const httpProxyMiddleware = () => middleware
+
+module.exports = httpProxyMiddleware