+++ config/constants.js
@@ -3,6 +3,8 @@ const { join } = require('path')
 // eslint-disable-next-line import/no-dynamic-require
 const FBS_CONF = require(join(process.cwd(), 'firebase.json'))
 
+const FRA_CONFIG = require('../helpers/getFraConfig')()
+
 const FUNCTIONS_DIR_NAME = FBS_CONF.functions
   ? FBS_CONF.functions.source
   : 'functions'
@@ -15,11 +17,18 @@ const SRC_DIR = join(process.cwd(), 'src')
 
 const CLIENT_ENTRY = join(SRC_DIR, 'client', 'index.jsx')
 
+const PORTS = {
+  emulator: FRA_CONFIG.emulatorPort || 5000,
+  devServer: FRA_CONFIG.devPort || 3000,
+}
+
 module.exports = {
+  FRA_CONFIG,
   FBS_CONF,
   CLIENT_ENTRY,
   FUNCTIONS_DIR_NAME,
   FUNCTIONS_DIR,
   HOSTING_DIR,
   SRC_DIR,
+  PORTS,
 }