+++ devServer/renderMiddleware.js
@@ -1,0 +1,3 @@
+const renderMiddleware = (_, res, next) => {
+
+}