+++ helpers/__tests__/deepenRelativePath.spec.js
@@ -1,4 +1,4 @@
-const deepenRelativePath = require('../../helpers/deepenRelativePath')
+const deepenRelativePath = require('../deepenRelativePath')
 
 describe('deepenRelativePath', () => {
   it('return same object if no package with local path were inside', () => {